+++ spec/acceptance/groups_spec.rb
@@ -188,9 +188,9 @@ resource "Groups" do
         expect(group.image.local_id).to eq localId
         expect(response_status).to eq(200)
         expect(json).to have_key("image")
-        expect(json["image"]).to have_key "clyid"
+        expect(json["image"]).to have_key "publicId"
         expect(json["image"]).to have_key "localId"
-        expect(json["image"]["clyid"]).to eq publicId
+        expect(json["image"]["publicId"]).to eq publicId
         expect(json["image"]["localId"]).to eq localId
       end
     end