+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - change type of media items duration from int to float
 - rename voice to audio
 - rename photo to image
+- rename clyid to publicId
 
 
 ## v0.5.3 on t2s2-dev